% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{summary.ame_mg}
\alias{summary.ame_mg}
\title{Summary of a multi group AME object}
\usage{
\method{summary}{ame_mg}(object, ...)
}
\arguments{
\item{object}{the output from an ameMG function}

\item{...}{unused}
}
\value{
a summary of parameter estimates and confidence intervals for a multi-group AME
}
\description{
Summary method for a multi group AME object. This is adapted from
Peter Hoff's work for ame class
}
\author{
Daniel Griffin
}
