% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{mgame}
\alias{mgame}
\title{Multi Group AME model}
\usage{
mgame(
  data,
  Y,
  group,
  Xdyad = NULL,
  Xego = NULL,
  Xalter = NULL,
  family,
  R = 0,
  rvar = !(family == "rrl"),
  cvar = TRUE,
  dcor = !symmetric,
  nvar = TRUE,
  intercept = !is.element(family, c("rrl", "ord")),
  symmetric = FALSE,
  seed = 1,
  nscan = 5000,
  burn = 500,
  odens = 25,
  odmax_grand = NULL,
  modelFitTest = TRUE,
  nullModel = TRUE,
  gof = TRUE,
  prior = list(),
  verboseOutput = FALSE,
  makePlot = FALSE
)
}
\arguments{
\item{data}{a data frame in person parwise structure}

\item{Y}{a dyadic outcome variable of interest}

\item{Xego}{an individual level anticedent for the ego node}

\item{Xalter}{an individual level anticedent for the alter node}

\item{family}{set to "nrm" for weighted dyadic outcom and "bin" for unweighted outcome values.
See ame documentation for more information.}

\item{modelFitTest}{weather to run fit test models}

\item{nullModel}{whether to run a null_model with only random additive
and multiplicitive effects to calculate unique varinace explained by fixed effects.}

\item{verboseOutput}{print ame model out put or not}

\item{makePlot}{generate plots as models are running or not}
}
\value{
a list of the class ame_mg which describes the posterior distribution
}
\description{
This function runs the ame model for a multi group setting
it also calculates fit statistics and variance explained by
the model.
}
